Feat: Edit styling and use svg image for the signup page
